Add Emerald.ApiHost API, routes, and services
Introduce a headless HTTP API host for Emerald CoreX: add new Emerald.ApiHost project files (routes, services registration, startup, request DTOs, and headless runtime settings) and refactor Program into partials. Implement REST endpoints for status, accounts, games, versions, runtime, settings, Java runtimes and notifications, plus a WebSocket event socket and background Core initialization. Configure Kestrel (including optional Unix socket), JSON enum serialization, Swagger/OpenAPI support and cleanup of stale socket files. Add package references for OpenAPI/Swashbuckle and update Directory.Packages.props accordingly.
