+++ Directory.Packages.props
@@ -14,6 +14,7 @@
     <PackageVersion Include="CommunityToolkit.WinUI.Helpers" Version="8.2.251219" />
     <PackageVersion Include="CommunityToolkit.WinUI.UI.Animations" Version="7.1.2" />
     <PackageVersion Include="CommunityToolkit.WinUI.UI.Controls" Version="7.1.2" />
+    <PackageVersion Include="Microsoft.AspNetCore.OpenApi" Version="10.0.8" />
     <PackageVersion Include="Microsoft.Windows.CsWin32" Version="0.3.269" />
     <PackageVersion Include="Microsoft.Windows.SDK.BuildTools" Version="10.0.26100.57" />
     <PackageVersion Include="Microsoft.WindowsAppSDK" Version="1.8.260317003" />
@@ -21,6 +22,7 @@
     <PackageVersion Include="Octokit" Version="13.0.1" />
     <PackageVersion Include="PInvoke.User32" Version="0.7.124" />
     <PackageVersion Include="ProjBobcat" Version="1.40.0" />
+    <PackageVersion Include="Swashbuckle.AspNetCore" Version="10.1.7" />
     <PackageVersion Include="System.Text.Json" Version="10.0.5" />
     <PackageVersion Include="WinUIEx" Version="2.3.4" />
     <PackageVersion Include="XboxAuthNet.Game.Msal" Version="0.1.3" />
