+++ Emerald.ApiHost/Emerald.ApiHost.csproj
@@ -9,4 +9,9 @@
   <ItemGroup>
     <ProjectReference Include="..\Emerald.CoreX\Emerald.CoreX.csproj" />
   </ItemGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.AspNetCore.OpenApi" />
+    <PackageReference Include="Swashbuckle.AspNetCore" />
+  </ItemGroup>
 </Project>
