+++ Emerald.ApiHost/HeadlessGameRuntimeSettings.cs
@@ -1,0 +1,8 @@
+using Emerald.CoreX.Runtime;
+
+namespace Emerald.ApiHost;
+
+public sealed class HeadlessGameRuntimeSettings : IGameRuntimeSettings
+{
+    public bool IsLogCaptureEnabled => true;
+}
